Move endpoint index from Decoder to CaptureShared.
